@@ -1,6 +1,7 @@
 from enum import Enum
+from typing import TypeAlias
 
-from typing_extensions import TypeAlias, reveal_type
+from typing_extensions import reveal_type
 
 from fastapi_filters import FilterField
 
@@ -56,31 +57,31 @@ reveal_type(field.not_contains(val))
 
 # output:
 """
-tests/mypy/enum_op.py:20:13: note: Revealed type is "fastapi_filters.op.FilterOp[enum.Enum]"
 tests/mypy/enum_op.py:21:13: note: Revealed type is "fastapi_filters.op.FilterOp[enum.Enum]"
-tests/mypy/enum_op.py:22:13: note: Revealed type is "None"
+tests/mypy/enum_op.py:22:13: note: Revealed type is "fastapi_filters.op.FilterOp[enum.Enum]"
 tests/mypy/enum_op.py:23:13: note: Revealed type is "None"
 tests/mypy/enum_op.py:24:13: note: Revealed type is "None"
 tests/mypy/enum_op.py:25:13: note: Revealed type is "None"
-tests/mypy/enum_op.py:28:13: note: Revealed type is "fastapi_filters.op.FilterOp[enum.Enum]"
+tests/mypy/enum_op.py:26:13: note: Revealed type is "None"
 tests/mypy/enum_op.py:29:13: note: Revealed type is "fastapi_filters.op.FilterOp[enum.Enum]"
-tests/mypy/enum_op.py:30:13: note: Revealed type is "None"
+tests/mypy/enum_op.py:30:13: note: Revealed type is "fastapi_filters.op.FilterOp[enum.Enum]"
 tests/mypy/enum_op.py:31:13: note: Revealed type is "None"
 tests/mypy/enum_op.py:32:13: note: Revealed type is "None"
 tests/mypy/enum_op.py:33:13: note: Revealed type is "None"
-tests/mypy/enum_op.py:36:13: note: Revealed type is "fastapi_filters.op.FilterOp[typing.Sequence[enum.Enum]]"
+tests/mypy/enum_op.py:34:13: note: Revealed type is "None"
 tests/mypy/enum_op.py:37:13: note: Revealed type is "fastapi_filters.op.FilterOp[typing.Sequence[enum.Enum]]"
 tests/mypy/enum_op.py:38:13: note: Revealed type is "fastapi_filters.op.FilterOp[typing.Sequence[enum.Enum]]"
-tests/mypy/enum_op.py:41:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.bool]"
+tests/mypy/enum_op.py:39:13: note: Revealed type is "fastapi_filters.op.FilterOp[typing.Sequence[enum.Enum]]"
 tests/mypy/enum_op.py:42:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.bool]"
 tests/mypy/enum_op.py:43:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.bool]"
-tests/mypy/enum_op.py:46:13: note: Revealed type is "Any"
+tests/mypy/enum_op.py:44:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.bool]"
 tests/mypy/enum_op.py:47:13: note: Revealed type is "Any"
 tests/mypy/enum_op.py:48:13: note: Revealed type is "Any"
 tests/mypy/enum_op.py:49:13: note: Revealed type is "Any"
-tests/mypy/enum_op.py:52:13: note: Revealed type is "None"
+tests/mypy/enum_op.py:50:13: note: Revealed type is "Any"
 tests/mypy/enum_op.py:53:13: note: Revealed type is "None"
 tests/mypy/enum_op.py:54:13: note: Revealed type is "None"
 tests/mypy/enum_op.py:55:13: note: Revealed type is "None"
+tests/mypy/enum_op.py:56:13: note: Revealed type is "None"
 Success: no issues found in 1 source file
 """
