@@ -1,4 +1,6 @@
-from typing_extensions import TypeAlias, reveal_type
+from typing import TypeAlias
+
+from typing_extensions import reveal_type
 
 from fastapi_filters import FilterField
 
@@ -47,31 +49,31 @@ reveal_type(field.not_contains(val))
 
 # output:
 """
-tests/mypy/str_op.py:11:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.str]"
-tests/mypy/str_op.py:12:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.str]"
-tests/mypy/str_op.py:13:13: note: Revealed type is "None"
-tests/mypy/str_op.py:14:13: note: Revealed type is "None"
+tests/mypy/str_op.py:13:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.str]"
+tests/mypy/str_op.py:14:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.str]"
 tests/mypy/str_op.py:15:13: note: Revealed type is "None"
 tests/mypy/str_op.py:16:13: note: Revealed type is "None"
-tests/mypy/str_op.py:19:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.str]"
-tests/mypy/str_op.py:20:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.str]"
-tests/mypy/str_op.py:21:13: note: Revealed type is "None"
-tests/mypy/str_op.py:22:13: note: Revealed type is "None"
+tests/mypy/str_op.py:17:13: note: Revealed type is "None"
+tests/mypy/str_op.py:18:13: note: Revealed type is "None"
+tests/mypy/str_op.py:21:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.str]"
+tests/mypy/str_op.py:22:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.str]"
 tests/mypy/str_op.py:23:13: note: Revealed type is "None"
 tests/mypy/str_op.py:24:13: note: Revealed type is "None"
-tests/mypy/str_op.py:27:13: note: Revealed type is "None"
-tests/mypy/str_op.py:28:13: note: Revealed type is "None"
+tests/mypy/str_op.py:25:13: note: Revealed type is "None"
+tests/mypy/str_op.py:26:13: note: Revealed type is "None"
 tests/mypy/str_op.py:29:13: note: Revealed type is "None"
-tests/mypy/str_op.py:32:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.bool]"
-tests/mypy/str_op.py:33:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.bool]"
+tests/mypy/str_op.py:30:13: note: Revealed type is "None"
+tests/mypy/str_op.py:31:13: note: Revealed type is "None"
 tests/mypy/str_op.py:34:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.bool]"
-tests/mypy/str_op.py:37:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.str]"
-tests/mypy/str_op.py:38:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.str]"
+tests/mypy/str_op.py:35:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.bool]"
+tests/mypy/str_op.py:36:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.bool]"
 tests/mypy/str_op.py:39:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.str]"
 tests/mypy/str_op.py:40:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.str]"
-tests/mypy/str_op.py:43:13: note: Revealed type is "None"
-tests/mypy/str_op.py:44:13: note: Revealed type is "None"
+tests/mypy/str_op.py:41:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.str]"
+tests/mypy/str_op.py:42:13: note: Revealed type is "fastapi_filters.op.FilterOp[builtins.str]"
 tests/mypy/str_op.py:45:13: note: Revealed type is "None"
 tests/mypy/str_op.py:46:13: note: Revealed type is "None"
+tests/mypy/str_op.py:47:13: note: Revealed type is "None"
+tests/mypy/str_op.py:48:13: note: Revealed type is "None"
 Success: no issues found in 1 source file
 """
